Replace deprecated altitude API with new JSON API
The XML API used for getting altitude/elevation data and
placeNumber/placeName has been shut down. This has broken the
`searchByCoordinates` function.

This commit replaces the altitude API with the new JSON API. The
"stedsnavn" API also must be used by `searchByCoordinates` now in order
to get `placeName` and `placeNumber`, so now there are 3 calls in
parallel for this function to work.

There is no need to parse XML any longer, so we can drop those
dependencies!

BREAKING:
- `searchByCoordinates` is using new APIs and might behave differently
from before.
- `placeNumber` is not actually `number`, not `string`.